+++ expat-journal/src/App.js
@@ -2,7 +2,8 @@ import React from 'react';
 import logo from './logo.svg';
 import './App.css';
 
-import Login from "../src/Components/Login/Login";
+// import Login from "../src/Components/Login/Login";
+import Login from "../src/Components/Login"
 
 function App() {
   return (
